rectangles are no longer overlapping :)
a rectangle is drawn initially, and then a new rectangle is drawn only when the pump status changes (e.g., one status change --> two rectangles). the x coordinate is the (scaled) time (milliseconds) at which the pump status changes. the width of the rectangle is the (scaled) number of milliseconds between the two start times.